+++ backends/tm_impl/tiny_stm/tinystm_wbetl.hpp
@@ -144,9 +144,6 @@ abort_tx(const char *loc = "")        //
 	if (tx->abort_count > 5) {
 		random_backoff();
 	}
-	if (g_tx_exit_jmpbuf && g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		siglongjmp(*g_tx_exit_jmpbuf, 1);
-	}
 	siglongjmp(*jmpbuf, 1);
 	TM_ASSERT(false, "Did not jump");
 }
@@ -154,9 +151,6 @@ abort_tx(const char *loc = "")        //
 inline bool //
 validate()  //
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	auto *tx = current_tx_wbetl;
 	for (auto &r : tx->read_set) {
 		void *addr = r.addr;
@@ -188,9 +182,6 @@ extend()    //
 inline bool //
 commit()    //
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	auto *tx = current_tx_wbetl;
 
 	TM_ASSERT(tx, "tx not defined");
@@ -251,9 +242,6 @@ read_word_etl(                                                //
     ValueType sz                                              //
 )
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	std::atomic_signal_fence(std::memory_order_seq_cst);
 	void *aligned = stm::merge::align_down_8(addr);
 	ByteOffset bo((word_t)aligned);
@@ -345,9 +333,6 @@ write_word_etl(                                               //
     ValueType sz                                              //
 )
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	std::atomic_signal_fence(std::memory_order_seq_cst);
 	void *aligned = stm::merge::align_down_8(addr);
 	ByteOffset bo((word_t)aligned);