+++ backends/tm_impl/tiny_stm/tinystm_wt.hpp
@@ -189,9 +189,6 @@ abort_tx(const char *loc = "") //
 		random_backoff();
 	}
 
-	if (g_tx_exit_jmpbuf && g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		siglongjmp(*g_tx_exit_jmpbuf, 1);
-	}
 	siglongjmp(*jmpbuf, 1);
 	TM_ASSERT(false, "Did not jump");
 }
@@ -200,9 +197,6 @@ abort_tx(const char *loc = "") //
 // Detects concurrent commits between read and write time.
 inline void validate_read_set_wt(word_t commit_version)
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	auto *tx = current_tx_wt;
 	if (commit_version <= tx->start_version + 1)
 		return;
@@ -247,9 +241,6 @@ inline void release_write_locks_wt(word_t commit_version)
 inline bool //
 commit()    //
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	auto *tx = current_tx_wt;
 
 	TM_ASSERT(tx, "commit: tx is null");
@@ -319,9 +310,6 @@ read_word_wt(                                           //
     ValueType /*sz*/                                    //
 )
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	std::atomic_signal_fence(std::memory_order_seq_cst);
 
 	TM_ASSERT(tx, "read_word_wt: tx is null");
@@ -438,9 +426,6 @@ write_word_wt(                                          //
     ValueType /*sz*/                                    //
 )
 {
-	if (g_tm_stop_requested.load(std::memory_order_relaxed)) {
-		abort_tx("proactive_stop"); // TODO.md: proactive_stop cleanup (P0)
-	}
 	std::atomic_signal_fence(std::memory_order_seq_cst);
 
 	TM_ASSERT(tx, "write_word_wt: tx is null");